refactor: split codebase-analyzer.ts into pipeline stage modules, remove dead RAG code
- Delete 1756-line server/services/codebase-analyzer.ts
- Add server/services/codebase-analyzer/ directory with:
  - constants.ts      — CODE_EXTENSIONS, SKIP_DIRS, DEPENDENCY_FILES + bug fix
  - file-discovery.ts — recursive file crawl (discoverFiles)
  - entity-extractor.ts — AST-lite parsing for 20+ languages
  - glossary.ts       — buildGlossary aggregator
  - doc-synthesizer.ts — generateDocumentation Markdown report
  - index.ts          — CodebaseAnalyzer orchestrator, public API re-exports
- Remove dead RAG code: createStructuredContent() and getMimeType() private
  methods that were unreachable after the ingestion phase was removed
- Fix *.csproj glob-in-Set bug: split DEPENDENCY_FILES (exact names) and
  DEPENDENCY_FILE_EXTENSIONS (.csproj, .fsproj) + isDependencyFile() helper;
  Set.has() does strict string equality so glob patterns never matched
- Existing callers (server/routes.ts) unchanged — dynamic import of
  './services/codebase-analyzer' resolves to the new index.ts via
  TypeScript bundler-mode directory index resolution

Closes #103
